Adds registration to Acorn application; calls from bootstrap.
Registration of configurable service providers is removed from `Acorn\Sage\SageServiceProvider` and added to `Acorn\Application`.

Adds a new bootstrap class to call this method: `Acorn\Bootstrap\RegisterProviders`. This bootstrap class has been added to the `$bootstrap` array in the `bootloader()` function of `src/helpers.php`.

With this change configurable service providers are successfully registered and booted based on the `providers` inventory in `config/app.php`.