Potential fix for broken view autoloading
In previous releases, you could forgo adding a `$views` prop if your
namespacing and class names reflected the structure of your views. It
looks (??) like a namespacing change at some point broke this, as I
wasn't able to get it to work on a recent project.

This commit fixes the issue, but is likely just a hack; consider this a
prompt for a better solution if mine is problematic.