make pest run test successfully on case insensitive filesystems
on a case insensitive filesystem, the functions defined in tests/helpers.php
were reported as being redeclared because pest (via phpunit) automatically
loads the that file. by renaming the file so that it will be found on
case-sensitive filesystems and removing the include from bootstrap.php,
things will run the same.